requirements.txt
langchain-groq
langchain
python-dotenv
ipykernel
langchain-community
langchain-ollama
langchain_huggingface
pypdf
faiss-cpu
chromadb
numexpr 
duckduckgo-search
wikipedia
playwright
unstructured
langgraph
jupyter
asyncio
langchain_experimental
pandas
matplotlib
graphviz
# API and Server
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
# Database
supabase>=2.0.0
redis>=5.0.1
# Utilities
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4